reorganize components into features folders
