+++ web/ui/src/pages/DAG.tsx
@@ -1,6 +1,6 @@
 import { faDiagramProject } from '@fortawesome/free-solid-svg-icons';
-import ComponentGraph from '../components/ComponentGraph';
-import Page from '../components/Page';
+import ComponentGraph from '../features/component/ComponentGraph';
+import Page from '../features/layout/Page';
 
 function DAG() {
   return (
