+++ web/ui/src/pages/status/BuildInfo.tsx
@@ -1,5 +1,5 @@
 import { faHammer } from '@fortawesome/free-solid-svg-icons';
-import Page from '../../components/Page';
+import Page from '../../features/layout/Page';
 
 function BuildInfo() {
   return <Page name="Build info" desc="Runtime and build information" icon={faHammer} />;
