+++ web/ui/src/pages/status/ConfigFile.tsx
@@ -1,5 +1,5 @@
 import { faGear } from '@fortawesome/free-solid-svg-icons';
-import Page from '../../components/Page';
+import Page from '../../features/layout/Page';
 
 function ConfigFile() {
   return <Page name="Config file" desc="Last successfully read unevaluated config file" icon={faGear} />;
