+++ web/ui/src/pages/status/Flags.tsx
@@ -1,5 +1,5 @@
 import { faFlag } from '@fortawesome/free-solid-svg-icons';
-import Page from '../../components/Page';
+import Page from '../../features/layout/Page';
 
 function Flags() {
   return <Page name="Flags" desc="Command-line flags used to launch Grafana Agent" icon={faFlag} />;
